+++ components/templates/Layouts/commonStyles.ts
