+++ components/templates/layouts/OtherLayout.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { commonLayoutStyle } from './style'
+import { commonLayoutStyle } from './commonStyles'
 
 const OtherLayout: React.FC = ({ children }) => {
   return <div css={commonLayoutStyle}>{children}</div>
